requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
a=1.0=pypi_0
absl-py=1.0.0=pypi_0
apptools=4.5.0=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.5=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37h89c1867_2
cffi=1.15.0=py37hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
configobj=5.0.6=pypi_0
cryptography=36.0.0=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
descartes=1.1.0=pypi_0
entrypoints=0.3=pypi_0
envisage=4.9.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filterpy=1.4.5=pypi_0
fire=0.4.0=pypi_0
freetype=2.11.0=h70c0345_0
freetype-py=2.1.0.post1=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
glfw=1.11.2=pypi_0
gmp=6.2.1=h2531618_2
gnureadline=8.0.0=pypi_0
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
idna=3.3=pyhd3eb1b0_0
imgui=1.1.0=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.2.1=pypi_0
ipython=7.14.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.0=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.3=pypi_0
jupyter-core=4.6.3=pypi_0
kiwisolver=1.2.0=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
miniball=1.1.0=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
msgpack-python=0.5.6=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.6=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=py37hd09550d_3
notebook=6.0.3=pypi_0
numpy=1.18.4=pypi_0
numpy-base=1.21.2=py37h79a1101_0
nuscenes-devkit=1.1.9=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
open3d=0.9.0.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1n=h7f8727e_0
packaging=20.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=pypi_0
pip=21.2.2=py37h06a4308_0
pointnet2=0.0.0=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=3.0.5=pypi_0
protobuf=3.20.0=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyface=7.0.0=pypi_0
pygments=2.6.1=pypi_0
pyopengl=3.1.5=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=2.4.7=pypi_0
pyqt5=5.14.0=pypi_0
pyqt5-sip=12.7.2=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.16.0=pypi_0
pysocks=1.7.1=py37_1
python=3.7.13=h12debd9_0
python-dateutil=2.8.1=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.10.2=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pyviz3d=0.2.28=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=58.0.4=py37h06a4308_0
shapely=1.8.1.post1=pypi_0
six=1.14.0=pypi_0
sqlite=3.38.2=hc218d9a_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.5.0=pypi_0
torchaudio=0.10.2=py37_cu113
torchdiffeq=0.1.0=pypi_0
torchvision=0.6.0=pypi_0
tornado=6.0.4=pypi_0
tqdm=4.48.2=pypi_0
traitlets=4.3.3=pypi_0
traits=6.0.0=pypi_0
traitsui=7.0.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
umsgpack=0.1.0=pypi_0
urllib3=1.26.8=pyhd3eb1b0_0
vispy=0.6.4=pypi_0
vtk=8.1.2=pypi_0
wcwidth=0.1.9=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
